server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic[email]
pydantic-settings
python-multipart
google-search-results
google-generativeai
langchain
langchain-core
langchain-google-genai
langgraph
motor
pymongo
passlib[bcrypt]
python-jose[cryptography]
email-validator
certifi